% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenList.R
\name{flattenList}
\alias{flattenList}
\title{A function to flatten a list of data frames to a single dataframe}
\usage{
flattenList(input_list)
}
\arguments{
\item{input_list}{as input list of data frames}
}
\value{
output_frame as data frame with rows indexed to original list position
}
\description{
A function to flatten a list of data frames to a single dataframe
}
\examples{
flattenList(nba_players)
}
